+++ src/lib.rs
@@ -1,3 +1,5 @@
+mod reference;
 pub mod schema;
 
+pub use self::reference::*;
 pub use self::schema::*;